Specify full index scan for using != as filter and rename the optimize rules
